Rework how create/variable work
Previously, these would need to emit a small chunk of code that pushed
the constant value on the stack, similar to jonesforth. The problem is
that this would break the common forth patterns:

   create myvar 10 cells allot
   create somearray 1 , 2 , 3 , 4 , 5

The create would assign the subsequent variable name to HERE, but that
would then point to the stub code. To remedy this, I added a new flag
to dictionary entries 'literal'. When these words are compiled, they
push a lit <value> inline instead of using a function stub.