Refactor how the palette is handled
On some places colors were stored as a packed integer, in others an array of
chars, which made the code clunky and confusing with all of the conversions.
Standardized in an array.
There was also a bug in sprite decoding that was masking the fact that
the indices stored in the files were wrong. I addressed that, but needed
to regenerate the examples.